Bug fixes for the brotli encoder.
  * Fix an out-of-bounds access to depth_histo in the
    bit cost calculation function.

  * Change type of distance symbol to uint16_t in block
    splitter, because if all postfix bits are used, there
    can be 520 distance symbols.

  * Save the distance cache between meta-blocks at the
    correct place. This fixes a roundtrip failure that
    can occur when there is an uncompressed metablock
    between two compressed metablocks.

  * Fix a bug when setting lgwin to 24 in the encoder parameters
    It ended up making metablocks larger than 24 bits in size.

  * Fix out-of-bounds memory accesses in parallel encoder.
    CreateBackwardReferences can read up to 4 bytes past end of
    input if the end of input is before mask.

  * Add missing header for memcpy() in port.h

+++ enc/block_splitter.cc
@@ -74,7 +74,7 @@ void CopyLiteralsToByteArray(const Command* cmds,
 void CopyCommandsToByteArray(const Command* cmds,
                              const size_t num_commands,
                              std::vector<uint16_t>* insert_and_copy_codes,
-                             std::vector<uint8_t>* distance_prefixes) {
+                             std::vector<uint16_t>* distance_prefixes) {
   for (int i = 0; i < num_commands; ++i) {
     const Command& cmd = cmds[i];
     insert_and_copy_codes->push_back(cmd.cmd_prefix_);
@@ -356,7 +356,7 @@ void SplitBlock(const Command* cmds,
 
   // Compute prefix codes for commands.
   std::vector<uint16_t> insert_and_copy_codes;
-  std::vector<uint8_t> distance_prefixes;
+  std::vector<uint16_t> distance_prefixes;
   CopyCommandsToByteArray(cmds, num_commands,
                           &insert_and_copy_codes,
                           &distance_prefixes);
